+++ src/command_modules/azure-cli-rdbms/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.2.5
++++++
+* Introduce mysql replica commands.
+
 0.2.4
 +++++
 * Minor fix allowing elastic server update using skuname.